refactor(examples-playwright): simplify setup
