+++ examples/with-playwright/.env.development
@@ -1,3 +1,3 @@
 NODE_ENV=development
 
-NEXT_PUBLIC_GITHUB_API_BASE_URL=https://api.github.com
+GITHUB_API_BASE_URL=https://api.github.com
