+++ examples/with-playwright/.env.test
@@ -1,4 +1,4 @@
 NODE_ENV=test
 
 ZIMIC_SERVER_URL=http://localhost:3003
-NEXT_PUBLIC_GITHUB_API_BASE_URL=$ZIMIC_SERVER_URL/github
+GITHUB_API_BASE_URL=$ZIMIC_SERVER_URL/github
