+++ examples/with-playwright/package.json
@@ -3,9 +3,10 @@
   "version": "0.0.0",
   "private": false,
   "scripts": {
-    "dev": "dotenv -c development -- next dev --port 3002",
-    "dev:test": "dotenv -c test -- pnpm dev",
-    "test": "dotenv -c test -- dotenv -c development -- zimic server start --port 3003 --ephemeral -- playwright test",
+    "dev": "dotenv -c development -- next dev --turbo --port 3002",
+    "dev:mock": "dotenv -c test -- zimic server start --port 3003 --ephemeral -- pnpm dev:load-interceptors -- pnpm dev",
+    "dev:load-interceptors": "tsx ./tests/interceptors/scripts/load.ts",
+    "test": "dotenv -c test -- dotenv -c development -- playwright test",
     "test:turbo": "pnpm run test",
     "types:check": "tsc --noEmit",
     "deps:install-playwright": "pnpm playwright install chromium",
@@ -25,6 +26,7 @@
     "@types/react-dom": "^18.3.0",
     "postcss": "^8.4.38",
     "tailwindcss": "^3.4.1",
+    "tsx": "^4.7.0",
     "typescript": "^5.4.3",
     "zimic": "latest"
   }
