+++ examples/with-playwright/src/app/__tests__/HomePage.e2e.test.ts
@@ -1,8 +1,6 @@
 import test, { expect } from '@playwright/test';
 
-import { githubFixtures } from '../../../tests/interceptors/github/fixtures';
-
-import '../../../tests/setup';
+import { githubFixtures } from '../../../tests/interceptors/github';
 
 test.describe('Home page', () => {
   const { repository } = githubFixtures;
