+++ examples/with-playwright/src/services/github.ts
@@ -3,7 +3,7 @@ import type { JSONValue } from 'zimic';
 
 import environment from '../config/environment';
 
-const BASE_URL = environment.GITHUB_API_BASE_URL;
+const GITHUB_API_BASE_URL = environment.GITHUB_API_BASE_URL;
 const CACHE_STRATEGY = process.env.NODE_ENV === 'production' ? 'default' : 'no-store';
 
 export type GitHubRepository = JSONValue<{
@@ -19,7 +19,7 @@ export const fetchGitHubRepository = cache(async (ownerName: string, repositoryN
     const sanitizedOwnerName = encodeURIComponent(ownerName);
     const sanitizedRepositoryName = encodeURIComponent(repositoryName);
 
-    const response = await fetch(`${BASE_URL}/repos/${sanitizedOwnerName}/${sanitizedRepositoryName}`, {
+    const response = await fetch(`${GITHUB_API_BASE_URL}/repos/${sanitizedOwnerName}/${sanitizedRepositoryName}`, {
       cache: CACHE_STRATEGY,
     });
